git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/proposed@18855 b456876b-0849-0410-b77d-98878d47e9d5

+++ doc/devtools/release.php
@@ -23,16 +23,24 @@ define( 'LICENSE_FILENAME', 'license.txt' );
 
 require_once TOOLS . '/svntools.php';
 
+if ( version_compare(PHP_VERSION, '5.0.0', '<') )
+	error("You need PHP version 5 or more to run this script\n");
+
+$phpCommand = isset($_SERVER['_']) ? $_SERVER['_'] : 'php';
+$phpCommandArguments = implode(' ', $_SERVER['argv']);
+
 if ( ! ( $options = get_options() ) || $options['help'] )
 	display_usage();
 
 if ( $options['howto'] )
 	display_howto();
 
+if ( ! check_svn_version() )
+	error("You need the subversion 'svn' program at least at version " . SVN_MIN_VERSION . "\n");
+
 if ( ! $options['no-check-svn'] && has_uncommited_changes('.') )
 	error("Uncommited changes exist in the working folder.\n");
 
-$phpCommand = isset($_SERVER['_']) ? $_SERVER['_'] : 'php';
 list( $script, $version, $subrelease ) = $_SERVER['argv'];
 
 if ( ! preg_match("/^\d+\.\d+$/", $version) )
@@ -46,6 +54,13 @@ if ( $isPre ) {
 	$pre = '';
 }
 $mainversion = $version{0} . '.0';
+
+include_once('lib/setup/twversion.class.php');
+$check_version = strtolower($version.$subrelease);
+$TWV = new TWVersion();
+if ( strtolower($TWV->version) != $check_version ) {
+	error("The version in the code ".strtolower($TWV->version)." differs from the version provided to the script $check_version.\nThe version should be modified in lib/setup/twversion.class.php to match the released version.");
+}
 
 echo color("\nTiki release process started for version '$version" . ( $subrelease ? " $subrelease" : '' ) . "'\n", 'cyan');
 if ( $isPre )
@@ -137,21 +152,24 @@ if ( $isPre ) {
 		echo color("\nMake sure these tarballs are tested by at least 3 different people.\n\n", 'cyan');
 	} else echo color("This was the last step.\n", 'cyan');
 } else {
-	$fb = full( $branch );
-	$ft = full( $tag );
 
-	$tagAlreadyExists = isset(get_info($ft)->entry);
-	if ( $tagAlreadyExists && important_step("The Tag '$tag' already exists: Delete the existing tag in order to create a new one") ) {
-		`svn rm $ft -m "[REL] Deleting tag '$tag' in order to create a new one"`;
-		$tagAlreadyExists = false;
-		info(">> Tag '$tag' deleted.");
-	}
+	if ( ! $options['no-tagging'] ) {
+		$fb = full( $branch );
+		$ft = full( $tag );
 
-	if ( ! $tagAlreadyExists ) {
-		$revision = (int) get_info( ROOT )->entry->commit['revision'];
-		if ( ! $options['no-tagging'] && important_step("Tag release using branch '$branch' at revision $revision") ) {
-			`svn copy $fb -r$revision $ft -m "[REL] Tagging release"`;
-			info(">> Tag '$tag' created.");
+		$tagAlreadyExists = isset(get_info($ft)->entry);
+		if ( $tagAlreadyExists && important_step("The Tag '$tag' already exists: Delete the existing tag in order to create a new one") ) {
+			`svn rm $ft -m "[REL] Deleting tag '$tag' in order to create a new one"`;
+			$tagAlreadyExists = false;
+			info(">> Tag '$tag' deleted.");
+		}
+
+		if ( ! $tagAlreadyExists ) {
+			$revision = (int) get_info( ROOT )->entry->commit['revision'];
+			if ( important_step("Tag release using branch '$branch' at revision $revision") ) {
+				`svn copy $fb -r$revision $ft -m "[REL] Tagging release"`;
+				info(">> Tag '$tag' created.");
+			}
 		}
 	}
 
@@ -164,9 +182,20 @@ if ( $isPre ) {
 // Helper functions
 
 function write_secdb( $file, $root, $version ) {
-	$fp = fopen( $file, 'w+' );
-	fwrite( $fp, "DELETE FROM `tiki_secdb` WHERE `tiki_version` = '$version';\n\n" );
-	md5_check_dir( $root, $root, $fp, $version );
+	if ( ! ($fp = @fopen($file, 'w+')) ) {
+		error('The SecDB file "' . $file . '" is not writable or can\'t be created.');
+		die;
+	}
+
+	$queries = array();
+	md5_check_dir( $root, $root, $version, $queries );
+
+	if ( ! empty($queries) ) {
+		sort($queries);
+		fwrite( $fp, "DELETE FROM `tiki_secdb` WHERE `tiki_version` = '$version';\n\n" );
+		foreach ( $queries as $q ) fwrite( $fp, "$q\n" );
+	}
+
 	fclose( $fp );
 
 	if ( $file_exists = file_exists($file) ) {
@@ -178,20 +207,33 @@ function write_secdb( $file, $root, $version ) {
 	}
 }
 
-function md5_check_dir($root,$dir,$fp,$version) { // save all files in $result
+function md5_check_dir($root, $dir, $version, &$queries) {
 	$d = dir($dir);
 	while (false !== ($e = $d->read())) {
 		$entry = $dir . '/' . $e;
 		if ( is_dir($entry) ) {
 			// do not descend and no CVS/Subversion files
 			if ( $e != '..' && $e != '.' && $e != 'CVS' && $e != '.svn' && $entry!='./templates_c') {
-				md5_check_dir($root, $entry, $fp, $version);
+				md5_check_dir($root, $entry, $version, $queries);
 			}
 		} else {
 			if ( substr($e, -4, 4) == ".php" && realpath( $entry ) != __FILE__ && $entry != './db/local.php' ) {
-				$file = mysql_real_escape_string( '.' . substr( $entry, strlen( $root ) ) );
+				$file = '.' . substr( $entry, strlen( $root ) );
+
+				if ( ! preg_match('/^[a-zA-Z0-9\/ _+.-]+$/', $file)
+					&& ( ! function_exists('mysql_real_escape_string') || ! ( $file = @mysql_real_escape_string($file) ) )
+				) {
+					global $phpCommand, $phpCommandArguments;
+					error("SecDB step failed because some filenames need escaping but no MySQL connection has been found."
+						. "\nTry this command line instead (replace HOST, USER and PASS by a valid MySQL host, user and password) :"
+						. "\n\n\t" . $phpCommand
+						. " -d mysql.default_host=HOST -d mysql.default_user=USER -d mysql.default_password=PASS "
+						. $phpCommandArguments . "\n"
+					);
+				}
+
 				$hash = md5_file($entry);
-				fwrite( $fp, "INSERT INTO `tiki_secdb` (`md5_value`, `filename`, `tiki_version`, `severity`) VALUES('$hash', '$file', '$version', 0);\n" );
+				$queries[] = "INSERT INTO `tiki_secdb` (`filename`, `md5_value`, `tiki_version`, `severity`) VALUES('$file', '$hash', '$version', 0);";
 			}
 		}
 	}
@@ -200,7 +242,7 @@ function md5_check_dir($root,$dir,$fp,$version) { // save all files in $result
 
 function build_packages($releaseVersion, $svnRelativePath) {
 	$script = TOOLS . '/tikirelease.sh';
-	`bash $script $releaseVersion $svnRelativePath`;
+	`bash -x "$script" $releaseVersion $svnRelativePath`;
 	info(">> Packages files have been built in ~/tikipack/$releaseVersion :\n");
 	passthru( "ls ~/tikipack/$releaseVersion" );
 }
@@ -268,6 +310,7 @@ function get_options() {
 		'howto' => false,
 		'help' => false,
 		'http-proxy' => false,
+		'no-commit' => false,
 		'no-check-svn' => false,
 		'no-check-php' => false,
 		'no-first-update' => false,
@@ -280,6 +323,18 @@ function get_options() {
 		'no-tagging' => false,
 		'force-yes' => false
 	);
+
+	// Environment variables provide default values for parameter options. e.g. export TIKI_NO_SECDB=true
+	$prefix = "TIKI-";
+	foreach ( $options as $option => $optValue) {
+	  $envOption = $prefix.$option;
+	  $envOption = str_replace("-", "_", $envOption);
+	  $envValue = $_ENV[$envOption];
+
+	  if (isset($envValue)) {
+	    $options[$option] = $envValue;
+	  }
+	}
 
 	foreach ( $_SERVER['argv'] as $arg ) {
 		if ( substr($arg, 0, 2) == '--' ) {
@@ -321,13 +376,27 @@ function important_step($msg, $increment_step = true, $commit_msg = false) {
 	// Increment step number if needed
 	if ( $increment_step ) $step++;
 
+	if ( $commit_msg && $options['no-commit'] ) {
+	  print "Skipping actual commit ('$commit_msg') because no-commit = true\n";
+	  return;
+	}
+
 	$do_step = false;
 	if ( $options['force-yes'] ) {
 		important("\n$step) $msg...");
 		$do_step = true;
 	} else {
 		important("\n$step) $msg?");
-		$c = readline('[Y/n/q/?] ');
+
+		$prompt = '[Y/n/q/?] ';
+		if ( function_exists('readline') ) {
+			// readline function requires php readline extension...
+			$c = readline($prompt);
+		} else {
+			echo $prompt;
+			$c = rtrim( fgets( STDIN ), "\n" );
+		}
+
 		switch ( strtolower($c) ) {
 			case 'y': case '':
 				$do_step = true;
@@ -346,8 +415,9 @@ function important_step($msg, $increment_step = true, $commit_msg = false) {
 		}
 	}
 
-	if ( $commit_msg && $do_step && $revision = commit($commit_msg) )
-		info(">> Commited revision $revision.");
+	if ( $commit_msg && $do_step ) {
+	  $revision = commit($commit_msg) && info(">> Commited revision $revision.");
+	}
 
 	return $do_step;
 }
@@ -602,6 +672,9 @@ function get_contributors_sf_data(&$contributors) {
 	$matches = array();
 	$userParsedInfo = array();
 
+	if ( ! function_exists('iconv') )
+		error("PHP 'iconv' function is not available on this system. Impossible to get SF.net data.");
+
 	$html = $options['http-proxy'] ? file_get_contents(SF_TW_MEMBERS_URL, 0, $options['http-proxy']) : file_get_contents(SF_TW_MEMBERS_URL);
 
 	if ( !empty($html) && preg_match('/(<table.*<\/\s*table>)/sim', $html, $matches) ) {
@@ -681,6 +754,7 @@ Options:
 	--howto			: display the Tiki release HOWTO
 	--help			: display this help
 	--http-proxy=HOST:PORT	: use an http proxy to get copyright data on sourceforge
+	--no-commit		: do not commit any changes back to SVN
 	--no-check-svn		: do not check if there is uncommited changes on the checkout used for the release
 	--no-check-php		: do not check syntax of all PHP files
 	--no-first-update	: do not svn update the checkout used for the release as the first step
